Moved socket communicator to it's own file since client_sdk is probably going to expand
